requirements.txt
astroid==2.4.1
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.1
mccabe==0.6.1
pylint==2.5.2
pymongo==3.10.1
python-dotenv==0.13.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
